requirements/saas.txt
# based on python 3.5.0

Django==1.11.4
django-admin-bootstrapped==2.5.4
django-angular==1.1
django-authtools==1.6.0
django-bootstrap3==9.0.0
django-braces==1.4.0
django-choices==1.5.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-envelope==1.3
django-environ==0.3.0
django-maintenance-mode==0.6.3
django-revproxy==0.9.13
django-storages-redux==1.3
djangorestframework==3.3.2
# django-ses
git+https://github.com/django-ses/django-ses.git@2fae0ebdaa96ff6daaea4edaafca3c02cdda61d8
easy-thumbnails==2.4.1
geojson==1.2.2
Jinja2==2.9.6
jsonfield==1.0.3
numpy==1.11.1
openpyxl==2.4.8
path.py==7.2
psycopg2
pymongo==3.3.0
python-dateutil==2.4.2
toposort==1.5

# For World Application
mapbox-vector-tile
mercantile
